Makefile now traverses src and added captains-log
The Makefile will now traverse the `src` directory and all its
subdirectories. It will populatae the `build` directory with all the
same directories that it finds in `src`.

This includes a new script, `captains-log`, which is intended to act
as a means of keeping track of daily activities. It maintains a
structured set of directories in `src/captains-log` for the year and
month with entries categorized in separate files for each day. The
same "day" is used up until an preset offset (currently 0400) into the
next day to try and better align with a real working schedule that
sometimes extends past midnight. When running this, it uses your
defined $EDITOR to open the log for the current day. When you close
the editor, this will populate a log of recent activity (the last 30
logs) and put this in a `captains-log/recent.md`. It will then run
make to regenerate anything that's changed.