Makefile looks in top-level directory for sources
In addition to looking in the src directory, the Makefile now looks in
the top-level directory. This is due to the way that GitHub handles
new pages of its browser-based wiki editing. All new pages show up in
the base of the repository (even though they are allowed to exist
anywhere in the repo). The Makefile will now build these and dump them
into build.